// src/app/core/api/models/app-schemas-infraestrutura-analise-temporal-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AnaliseTemporalData } from '../models/analise-temporal-data';

/**
 * Resposta da análise temporal por tipo de localização e dependência.
 */
export interface AppSchemasInfraestruturaAnaliseTemporalResponse {
  data: AnaliseTemporalData;

  /**
   * Identificador semântico do recurso.
   */
  descricao: string;
}