// src/app/core/api/models/app-schemas-infraestrutura-mapa-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AppSchemasInfraestruturaMapaData } from '../models/app-schemas-infraestrutura-mapa-data';

/**
 * Resposta do mapa: descrição + lista de pontos com score/classificação.
 */
export interface AppSchemasInfraestruturaMapaResponse {
  data: AppSchemasInfraestruturaMapaData;

  /**
   * Identificador semântico do mapa.
   */
  descricao: string;
}